Create .travis.yml
Boldly taken from https://github.com/davidwhogg/mcmc/blob/master/.travis.yml

Here's to hoping I didn't break anything. We shouldn't need most of this (currently no python) but if there is one language we are likely to add in the future for making plots, it is python, so I am leaving it in.